add back-face occlusion

+++ src/Terrain.cpp
@@ -1,7 +1,7 @@
 #include "Terrain.hpp"
 #include "glm/ext.hpp"
 
-Terrain::Terrain( uint8_t renderDistance, uint8_t maxHeight, const glm::ivec3& chunkSize ) : renderDistance(renderDistance), maxHeight(maxHeight), chunkSize(chunkSize) {
+Terrain::Terrain( uint8_t renderDistance, uint maxHeight, const glm::ivec3& chunkSize ) : renderDistance(renderDistance), maxHeight(maxHeight), chunkSize(chunkSize) {
     this->setupChunkGenerationRenderingQuad();
     this->setupChunkGenerationFbo();
     this->chunkGenerationShader = new Shader("./shader/vertex/screenQuad.vert.glsl", "./shader/fragment/generateChunk.frag.glsl");
